+++ Classes/xpopup/xpopupmenu.h
@@ -145,9 +145,9 @@ public:
 	void setBackBitmap( HBITMAP hBitmap );
 
 	bool IsRounded(void) const { return this->m_bRoundedSel; };
-	UINT IsAlpha(void) const { return this->m_uiAlpha; };
+	BYTE IsAlpha(void) const { return this->m_uiAlpha; };
 	void SetRounded(bool rounded) { this->m_bRoundedSel = rounded; };
-	void SetAlpha(UINT alpha) { this->m_uiAlpha = alpha; };
+	void SetAlpha(BYTE alpha) { this->m_uiAlpha = alpha; };
 
 	// Methods to attach and detach from mIRC menu.
 	bool attachToMenuBar(HMENU menubar, TString label);
@@ -173,7 +173,7 @@ protected:
 	XPMENUCOLORS m_MenuColors; //!< Menu Colors
 
 	bool m_bRoundedSel; //!< Menu has rounded selection box.
-	UINT m_uiAlpha;			//!< Menu is alpha blended. 0 -> 255
+	BYTE m_uiAlpha;			//!< Menu is alpha blended. 0 -> 255
 
 	bool m_bAttachedToMenuBar; //!< Is the menu attached to the mIRC window menubar?
 };
